walletApp/requirements.txt
arrow==1.2.2
asgiref==3.5.2
blessed==1.19.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.9
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-filter==22.1
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.13.1
drf-writable-nested==0.6.3
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
psycopg2-binary==2.9.1
pycryptodome==3.15.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==3.5.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
wcwidth==0.2.5
celery==4.4.7
redis==3.5.3
drf_yasg